Add files via upload
TOKYO
Fast food and eating clean are two phrases that shouldn’t really be seen side by side, but nevertheless, this new store in Osaka is proving they can go hand in hand.

Since the outbreak of COVID-19 and the following state of emergencies, individuals have watched their health deteriorate as a result of a lack of exercise and narrow food choices. As expected these issues have led to a rapid increase in the interest of health.